* rebrand-devel:
  branding: update assembly information
  Product.wxs: change to per-user install
  Product.wxs: avoid hard coded paths in registry
  Product.wxs: reset options back to defaults
  Product.wxs: fix generated noise commit
  update base domain
  Logger: read before write crash
  SetupHelp.cs: check directory before removing
  Console: exit on thrown exception
  Shared.modules: update register path
  update sign script
  Addin.Setup: drop install shield project
  UDF: update prompt branding
  README: add wix editor extension
  Logger: update log path
  update installer copy and branding
  Settings: remove references to quandl
  rebrand updates
  Addin.Setup: update link in EULA

+++ Quandl.Excel.Console/SetupHelp.cs
@@ -149,8 +149,11 @@ namespace Quandl.Excel.Console
 
         private static void CleanLogs()
         {
-            var documentsPath = Path.Combine(Environment.ExpandEnvironmentVariables("%userprofile%"), "Documents", "Quandl");
-            Directory.Delete(documentsPath, true);
+            var documentsPath = Path.Combine(Environment.ExpandEnvironmentVariables("%userprofile%"), "Documents", "Nasdaq");
+            if (Directory.Exists(documentsPath))
+            {
+                Directory.Delete(documentsPath, true);
+            }
         }
